unify all file source tests
