+++ ci/workflows/main-cron.yml
@@ -486,28 +486,6 @@ steps:
     timeout_in_minutes: 25
     retry: *auto-retry
 
-  - label: "S3 source new file check on AWS (json)"
-    key: "s3-v2-source-new-file-check-aws"
-    command: "ci/scripts/s3-source-test.sh -p ci-release -s fs_source_v2_new_file.py"
-    if: |
-      !(build.pull_request.labels includes "ci/main-cron/run-selected") && build.env("CI_STEPS") == null
-      || build.pull_request.labels includes "ci/run-s3-source-tests"
-      || build.env("CI_STEPS") =~ /(^|,)s3-source-tests?(,|$$)/
-    depends_on: build
-    plugins:
-      - seek-oss/aws-sm#v2.3.1:
-          env:
-            S3_SOURCE_TEST_CONF: ci_s3_source_test_aws
-      - docker-compose#v5.1.0:
-          run: rw-build-env
-          config: ci/docker-compose.yml
-          mount-buildkite-agent: true
-          environment:
-            - S3_SOURCE_TEST_CONF
-      - ./ci/plugins/upload-failure-logs
-    timeout_in_minutes: 25
-    retry: *auto-retry
-
   - label: "S3 sink on parquet and json file"
     key: "s3-sink-parquet-and-json-encode"
     command: "ci/scripts/s3-source-test.sh -p ci-release -s file_sink.py"
